+++ packages/react-moveable/src/react-moveable/InitialMoveable.tsx
@@ -3,12 +3,14 @@ import { Able, MoveableInterface, GroupableProps, MoveableDefaultProps } from ".
 import MoveableManager from "./MoveableManager";
 import MoveableGroup from "./MoveableGroup";
 import { ref, withMethods, prefixCSS } from "framework-utils";
-import { isArray, getKeys, IObject } from "@daybrush/utils";
+import { getKeys, IObject } from "@daybrush/utils";
 import { MOVEABLE_METHODS, PREFIX, MOVEABLE_CSS } from "./consts";
 import Default from "./ables/Default";
 import Groupable from "./ables/Groupable";
 import DragArea from "./ables/DragArea";
 import styled from "react-css-styled";
+import { getRefTargets, getElementTargets } from "./utils";
+import { isString } from "util";
 
 export class InitialMoveable<T = {}>
     extends React.PureComponent<MoveableDefaultProps & GroupableProps & T> {
@@ -36,16 +38,20 @@ export class InitialMoveable<T = {}>
     }
     @withMethods(MOVEABLE_METHODS)
     public moveable!: MoveableManager | MoveableGroup;
-
+    public refTargets: Array<HTMLElement | SVGElement | string | undefined | null> = [];
+    public selectorMap: IObject<Array<HTMLElement | SVGElement>> = {};
     public render() {
         const moveableContructor = (this.constructor as typeof InitialMoveable);
 
         if (!moveableContructor.defaultStyled) {
             moveableContructor.makeStyled();
         }
-        const target = this.props.target || this.props.targets;
-        const isArr = isArray(target);
-        const isGroup = isArr && (target as any[]).length > 1;
+        const refTargets = getRefTargets((this.props.target || this.props.targets) as any);
+        const elementTargets = getElementTargets(refTargets, this.selectorMap);
+
+        this.refTargets = refTargets;
+
+        const isGroup = elementTargets.length > 1;
         const totalAbles = moveableContructor.getTotalAbles();
         const userAbles = this.props.ables! || [];
         const ables = [
@@ -54,21 +60,55 @@ export class InitialMoveable<T = {}>
         ];
 
         if (isGroup) {
-            const nextProps = {
-                ...this.props,
-                target: null,
-                targets: target as any[],
-                ables,
-            } as any;
             return <MoveableGroup key="group" ref={ref(this, "moveable")}
                 cssStyled={moveableContructor.defaultStyled}
-                {...nextProps} />;
+                {...this.props}
+                target={null}
+                targets={elementTargets}
+                ables={ables} />;
         } else {
-            const moveableTarget = isArr ? (target as any[])[0] : target;
-
             return <MoveableManager<any> key="single" ref={ref(this, "moveable")}
                 cssStyled={moveableContructor.defaultStyled}
-                {...{ ...this.props, target: moveableTarget, ables }} />;
+                {...this.props}
+                target={elementTargets[0]}
+                ables={ables} />;
+        }
+    }
+    public componentDidMount() {
+        this.updateRefs();
+    }
+    public componentDidUpdate() {
+        this.updateRefs();
+    }
+    public updateRefs(isReset?: boolean) {
+        const refTargets = getRefTargets((this.props.target || this.props.targets) as any);
+        let isUpdate = this.refTargets.some((target, i) => {
+            const nextTarget = refTargets[i];
+
+            if (!target && !nextTarget) {
+                return false;
+            } else if (target !== nextTarget) {
+                return true;
+            }
+            return false;
+        });
+        const selectorMap = isReset ? {} : this.selectorMap;
+        const nextSelectorMap: IObject<Array<HTMLElement | SVGElement>> = {};
+        this.refTargets.forEach(target => {
+            if (isString(target)) {
+                if (!selectorMap[target]) {
+                    isUpdate = true;
+                    nextSelectorMap[target] = [].slice.call(document.querySelectorAll(target));
+                } else {
+                    nextSelectorMap[target] = selectorMap[target];
+                }
+            }
+        });
+
+        this.selectorMap = nextSelectorMap;
+
+        if (isUpdate) {
+            this.forceUpdate();
         }
     }
 }