+++ packages/react-moveable/src/react-moveable/makeMoveable.ts
@@ -1,7 +1,8 @@
 import { Able } from "./types";
 import { InitialMoveable } from "./InitialMoveable";
 
-export function makeMoveable<T = {}>(ables: Array<Able<T>>) {
+export function makeMoveable<T = {}>(
+    ables: Array<Able<T>>): typeof InitialMoveable & (new (...args: any[]) => InitialMoveable<T>) {
     return class Moveable extends InitialMoveable<T> {
         public static defaultAbles = ables;
     };